:hammer: moved class json_pointer into separate file #920
